+++ admin/components/head.admin.php
@@ -9,9 +9,12 @@ $assetService = AssetService::getInstance();
 <html>
 <head>
     <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0 user-scalable=no">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0 user-scalable=no, viewport-fit=cover">
+    <meta name="mobile-web-app-capable" content="yes" />
     <meta name="msapplication-TileColor" content="<?=$config['colors']['primary']?>">
-    <meta name="theme-color" content="<?=$config['colors']['primary']?>">
+    <meta name="theme-color" content="<?=$config['colors']['status_bar'] ?? '#000000'?>">
+    <meta name="apple-mobile-web-app-capable" content="yes" />
+    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
 
     <title><?=$pageTitle ?></title>