+++ lib/boot.php
@@ -31,6 +31,7 @@ FileUtility::createDirectory(FolderEnum::THUMBS->absolute());
 FileUtility::createDirectory(FolderEnum::TEMP->absolute());
 FileUtility::createDirectory(FolderEnum::PRIVATE->absolute());
 FileUtility::createDirectory(PathUtility::getAbsolutePath('private/fonts'));
+FileUtility::createDirectory(PathUtility::getAbsolutePath('private/screensavers'));
 FileUtility::createDirectory(PathUtility::getAbsolutePath('private/images/background'));
 FileUtility::createDirectory(PathUtility::getAbsolutePath('private/images/frames'));
 FileUtility::createDirectory(PathUtility::getAbsolutePath('private/images/keyingBackgrounds'));